+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to this project will be documented in this file.
 
+## 2026-09-03
+
+### Fixed
+- **Historical sliding scales apply properly to past programs and balances**: When viewing past programs or historical balance sheets, sliding scales active during the program's timeframe now correctly discount applicable program fees even if the sliding scale has an expiration date in the past.
+
 ## 2026-09-02
 
 ### Added